fix(physical-backup): Resolve cyclic imports
